+++ src/main/java/eu/fthevenet/binjr/sources/adapters/elitebgs/api/v4/EBGSFactionPresenceV4.java
@@ -17,4 +17,10 @@
 package eu.fthevenet.binjr.sources.adapters.elitebgs.api.v4;
 
 public class EBGSFactionPresenceV4 {
+    public String system_name;
+    public String system_name_lower;
+    public String state;
+    public double influence;
+    public EBGSStateV4[] pending_states;
+    public EBGSStateV4[] recovering_states;
 }