+++ CHANGELOG.md
@@ -1,1 +1,4 @@
-Preview v0.1.0
+## [Preview v0.1.0](https://github.com/fthevenet/binjr-adapter-elitebgs/releases/tag/v0.1.0)
+Released on Sun, 18 Aug 2019
+
+* First public preview